✨ Admin api

+++ src/modules/admin/affiliates.controller.ts
@@ -10,7 +10,7 @@ import {
   Post,
   UseGuards,
 } from '@nestjs/common';
-import { AffiliateService } from './affiliate.service';
+import { AffiliateService } from '../affiliate/affiliate.service';
 import { ApiKeyGuard } from '@/decorators/api-key.decorator';
 import { ApiBearerAuth } from '@nestjs/swagger';
 import {
@@ -20,17 +20,21 @@ import {
 } from '@/types';
 import { errorMessage } from '@/errors';
 import { affiliateValidation } from '@/validations';
+import { GetCurrentUser } from '@/decorators/get-current-user.decorator';
+import { User } from '../user/user.entity';
 
-@Controller('affiliates')
-export class AffiliateController {
+@Controller('admin')
+export class AdminAffiliatesController {
   constructor(private readonly service: AffiliateService) {}
 
-  @Get()
+  @Get('affiliates')
   @HttpCode(200)
   @UseGuards(ApiKeyGuard)
   @ApiBearerAuth()
-  async getAffiliates() {
+  async getAffiliates(@GetCurrentUser() user: User) {
     try {
+      if (!user.isAdmin)
+        throw new BadRequestException(errorMessage.api('admin').NOT_ADMIN);
       const affiliates = await this.service.getAffiliates();
       return affiliates.map((affiliate) =>
         this.service.formatAffiliate(affiliate),
@@ -40,14 +44,17 @@ export class AffiliateController {
     }
   }
 
-  @Get(':animalType')
+  @Get('affiliates/:animalType')
   @HttpCode(200)
   @UseGuards(ApiKeyGuard)
   @ApiBearerAuth()
   async getAffiliateByAnimalType(
     @Param('animalType') animalType: AnimalTypeEnum,
+    @GetCurrentUser() user: User,
   ) {
     try {
+      if (!user.isAdmin)
+        throw new BadRequestException(errorMessage.api('admin').NOT_ADMIN);
       if (AnimalTypeEnum[animalType] === undefined)
         throw new BadRequestException(
           errorMessage.api('billingPlan').NOT_FOUND,
@@ -62,12 +69,17 @@ export class AffiliateController {
     }
   }
 
-  @Post()
+  @Post('affiliates')
   @HttpCode(201)
   @UseGuards(ApiKeyGuard)
   @ApiBearerAuth()
-  async createBillingPlan(@Body() body: CreateAffiliateApi) {
+  async createBillingPlan(
+    @Body() body: CreateAffiliateApi,
+    @GetCurrentUser() user: User,
+  ) {
     try {
+      if (!user.isAdmin)
+        throw new BadRequestException(errorMessage.api('admin').NOT_ADMIN);
       await affiliateValidation.create.validate(body, {
         abortEarly: false,
       });
@@ -78,15 +90,18 @@ export class AffiliateController {
     }
   }
 
-  @Patch(':id')
+  @Patch('affiliates/:id')
   @HttpCode(200)
   @UseGuards(ApiKeyGuard)
   @ApiBearerAuth()
   async updateBillingPlan(
     @Param('id') id: string,
     @Body() body: UpdateAffiliateApi,
+    @GetCurrentUser() user: User,
   ) {
     try {
+      if (!user.isAdmin)
+        throw new BadRequestException(errorMessage.api('admin').NOT_ADMIN);
       await affiliateValidation.update.validate(body, {
         abortEarly: false,
       });
@@ -97,12 +112,17 @@ export class AffiliateController {
     }
   }
 
-  @Delete(':id')
+  @Delete('affiliates/:id')
   @HttpCode(200)
   @UseGuards(ApiKeyGuard)
   @ApiBearerAuth()
-  async deleteBillingPlan(@Param('id') id: string) {
+  async deleteBillingPlan(
+    @Param('id') id: string,
+    @GetCurrentUser() user: User,
+  ) {
     try {
+      if (!user.isAdmin)
+        throw new BadRequestException(errorMessage.api('admin').NOT_ADMIN);
       await this.service.deleteAffiliate(id);
     } catch (e) {
       throw new BadRequestException(e);
